requirements.txt
# MindSentinel - Dependencias del Proyecto
# ==========================================

# --- Deep Learning ---
tensorflow
keras

# --- Data Science ---
numpy
pandas
scikit-learn

# --- Visualización ---
matplotlib
seaborn

# --- NLP ---
nltk

# --- Frontend ---
streamlit

# --- Multi-Agente y LLM ---
crewai
crewai-tools
langchain
langchain-google-genai
google-generativeai
crewai[google-genai]
litellm

# --- Utilidades ---
python-dotenv
kagglehub